requirements.txt
anyio==3.3.4
asgiref==3.4.1
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
ecdsa==0.17.0
fastapi==0.70.0
h11==0.12.0
idna==3.3
mysql-connector-python==8.0.27
passlib==1.7.4
protobuf==3.19.1
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
python-decouple==3.5
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
